rihab commit4

+++ database/migrations/2025_11_16_172033_add_verrouiller_to_users_table.php
@@ -6,17 +6,22 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    public function up(): void
-    {
-        Schema::table('users', function (Blueprint $table) {
-            $table->boolean('verrouiller')->default(0)->after('adresse'); 
-        });
-    }
+public function up(): void
+{
+    Schema::table('users', function (Blueprint $table) {
+        if (!Schema::hasColumn('users', 'verrouiller')) {
+            $table->boolean('verrouiller')->default(0)->after('adresse');
+        }
+    });
+}
 
-    public function down(): void
-    {
-        Schema::table('users', function (Blueprint $table) {
+public function down(): void
+{
+    Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasColumn('users', 'verrouiller')) {
             $table->dropColumn('verrouiller');
-        });
-    }
+        }
+    });
+}
+
 };
